+++ amoro-ams/pom.xml
@@ -203,7 +203,6 @@
         <dependency>
             <groupId>mysql</groupId>
             <artifactId>mysql-connector-java</artifactId>
-            <scope>provided</scope>
         </dependency>
 
         <dependency>
